Run inferCNV sample by sample
inferCNV is a computationally intensive program to run and parameters need to be optimized and preferably run sample by sample to reduce overload and improve both speed and efficiency. Here a general R script is provided to run the inferCNV pipeline. However, each sample and parameters to be used should be assessed on a case by case basis.