General inferCNV parameters done sample by sample
Here we provide a general script to run the inferCNV pipeline. inferCNV is a computationally intensive program to run and optimizing parameters sample by sample is recommended to improve speed and efficiency of the pipeline. Parameters may need to be modified for each sample and this should be assessed on a case by case basis.